# flulink/backend/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FluLink API",
    "version": "1.0.0",
    "description": "FluLink星尘共鸣版API文档",
    "contact": {
      "name": "FluLink Team",
      "email": "support@flulink.com",
      "url": "https://flulink.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "开发环境"
    },
    {
      "url": "https://flulink-app.zeabur.app",
      "description": "生产环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "错误信息"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/resonance/calculate": {
      "post": {
        "summary": "计算共鸣值",
        "description": "计算两个用户之间的共鸣强度",
        "tags": [
          "Resonance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "targetUserId"
                ],
                "properties": {
                  "targetUserId": {
                    "type": "string",
                    "description": "目标用户ID",
                    "example": "60f7b3b3b3b3b3b3b3b3b3b3"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "共鸣值计算成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "resonance": {
                          "$ref": "#/components/schemas/Resonance"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未授权访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}